+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export { Code } from './Code';
 export { ContentEditable } from './ContentEditable';
 export { History } from './History';
 export { Selection } from './Selection';
+export { BlockNavigation } from './BlockNavigation';